fix: use correct ValidityType data type per IPNS specification
## Issue
ValidityType was using i32 instead of the required u64 type specified
in the IPNS V2 record format specification.

## Fix
Change ValidityType from i32 to u64 in SignatureV2Data CBOR structure
to comply with IPFS IPNS specification requirements.

## Verification
✅ Create new IPNS records works
✅ Update existing IPNS records works
✅ CBOR data structure now matches IPNS spec

+++ w3name/src/ipns/mod.rs
@@ -140,7 +140,7 @@ fn validate_v2_data_matches_entry_data(
     || entry.validity != data.Validity
     || entry.sequence != data.Sequence
     || entry.ttl != data.TTL
-    || entry.validity_type != data.ValidityType
+    || entry.validity_type != data.ValidityType as i32
   {
     Err(report!(InvalidIpnsV2SignatureData))
   } else {
@@ -184,7 +184,7 @@ struct SignatureV2Data {
   Value: Vec<u8>,
   #[serde(with = "serde_bytes")]
   Validity: Vec<u8>,
-  ValidityType: i32,
+  ValidityType: u64,
   Sequence: u64,
   TTL: u64,
 }
